position 3 saniyede bir değiştirildi

+++ app/src/main/java/com/example/satellitesfinder/model/Positions.kt
@@ -10,6 +10,6 @@ class SingleSatellite() {
 }
 
 class PositionsOfSatellite() {
-    var posX: Float? = null
-    var posY: Float? = null
+    var posX: Double? = null
+    var posY: Double? = null
 }
